requirements.txt
pyTelegramBotAPI
requests
Pillow
Flask